Add a database connection manager.
This is comprised of a singleton abstract class that can be called
to get the connection suitable for a given orm class.

This commit provides a singleton implementation, which simply wraps
the existing database singleton setup, and a multiple connection
manager that allows each orm class to have its own connection.