Have the ormCollection do its own constructing.
Move the construction of the orm objects that make up the results
into the collection's constructor. This is less than ideal, since
we need to pass the name in, but it's better than the internals
(the result's properties' naming convention) being used outside the
class.

This might work better as an addResult method or similar.